backend/src/requirements.txt
requests
pydantic
boto3
pydantic[email]
email-validator
python-jose